Define a class called PackageAPI and put it in its own file.
Summary:
The methods of the ad-hoc `api` object needed access to very few variables
from the surrounding lexical environment, and those few variables
(`sources`, `exports`, `uses`, `implies`) are better regarded as member
variables of the `PackageAPI` class.

Test Plan: `meteor test-packages`

Reviewers: Slava, dgreensp, glasser

Reviewed By: glasser

Differential Revision: https://phabricator.meteor.io/D30
