+++ tools/package-source.js
@@ -13,6 +13,7 @@ var release = require('./release.js');
 var catalog = require('./catalog.js');
 var packageVersionParser = require('./package-version-parser.js');
 var compiler = require('./compiler.js');
+var packageAPIModule = require('./package-api.js');
 
 // XXX: This is a medium-term hack, to avoid having the user set a package name
 // & test-name in package.describe. We will change this in the new control file
@@ -76,20 +77,6 @@ var loadOrderSort = function (templateExtensions) {
     // otherwise alphabetical
     return (a < b ? -1 : 1);
   };
-};
-
-// We currently have a 1 to 1 mapping between 'where' and 'arch'.
-// 'client' -> 'web'
-// 'server' -> 'os'
-// '*' -> '*'
-var mapWhereToArch = function (where) {
-  if (where === 'server') {
-    return 'os';
-  } else if (where === 'client') {
-    return 'web';
-  } else {
-    return where;
-  }
 };
 
 var splitConstraint = function (c) {
@@ -475,7 +462,6 @@ _.extend(PackageSource.prototype, {
     var code = files.readFile(packageJsPath);
     var packageJsHash = watch.sha1(code);
 
-    var releaseRecords = [];
     var hasTests = false;
 
     // Any package that depends on us needs to be rebuilt if our package.js file
@@ -588,7 +574,7 @@ _.extend(PackageSource.prototype, {
        * @summary Define package dependencies and expose package methods.
        * @locus package.js
        * @memberOf Package
-       * @param {Function} func A function that takes in the package control 'api' object, which keeps track of dependencies and exports.
+       * @param {Function} func A function that takes in the package control `api` object, which keeps track of dependencies and exports.
        */
       onUse: function (f) {
         if (!self.isTest) {
@@ -958,37 +944,6 @@ _.extend(PackageSource.prototype, {
       // recover by ignoring
     }
 
-    // source files used
-    var sources = {};
-
-    // symbols exported
-    var exports = {};
-
-    // packages used and implied (keys are 'package', 'unordered', and
-    // 'weak').  an "implied" package is a package that will be used by a unibuild
-    // which uses us.
-    var uses = {};
-    var implies = {};
-
-    _.each(compiler.ALL_ARCHES, function (arch) {
-      sources[arch] = [];
-      exports[arch] = [];
-      uses[arch] = [];
-      implies[arch] = [];
-    });
-
-    // Iterates over the list of target archs and calls f(arch) for all archs
-    // that match an element of self.allarchs.
-    var forAllMatchingArchs = function (archs, f) {
-      _.each(archs, function (arch) {
-        _.each(compiler.ALL_ARCHES, function (matchArch) {
-          if (archinfo.matches(matchArch, arch)) {
-            f(matchArch);
-          }
-        });
-      });
-    };
-
     // For this old-style, onUse/onTest/where-based package, figure
     // out its dependencies by calling its on_xxx functions and seeing
     // what it does.
@@ -1002,348 +957,11 @@ _.extend(PackageSource.prototype, {
     // probably is sufficient for virtually all packages that actually
     // exist in the field, if not every single one. #OldStylePackageSupport
 
+    var api = new packageAPIModule.PackageAPI({
+      buildingIsopackets: !!initFromPackageDirOptions.buildingIsopackets
+    });
+
     if (fileAndDepLoader) {
-      var toArray = function (x) {
-        if (x instanceof Array)
-          return x;
-        return x ? [x] : [];
-      };
-
-      var toArchArray = function (arch) {
-        if (!(arch instanceof Array)) {
-          arch = arch ? [arch] : compiler.ALL_ARCHES;
-        }
-        arch = _.uniq(arch);
-        arch = _.map(arch, mapWhereToArch);
-        _.each(arch, function (inputArch) {
-          var isMatch = _.any(_.map(compiler.ALL_ARCHES, function (actualArch) {
-            return archinfo.matches(actualArch, inputArch);
-          }));
-          if (! isMatch) {
-            buildmessage.error(
-              "Invalid 'where' argument: '" + inputArch + "'",
-              // skip toArchArray in addition to the actual API function
-              {useMyCaller: 2});
-          }
-        });
-        return arch;
-      };
-
-      /**
-       * @class PackageAPI
-       * @instanceName api
-       * @global
-       * @summary The API object passed into the Packages.onUse function.
-       */
-      var api = {
-        // Called when this package wants to make another package be
-        // used. Can also take literal package objects, if you have
-        // anonymous packages you want to use (eg, app packages)
-        //
-        // @param arch 'web', 'web.browser', 'web.cordova', 'server',
-        // or an array of those.
-        // The default is ['web', 'server'].
-        //
-        // options can include:
-        //
-        // - unordered: if true, don't require this package to load
-        //   before us -- just require it to be loaded anytime. Also
-        //   don't bring this package's imports into our
-        //   namespace. If false, override a true value specified in
-        //   a previous call to use for this package name. (A
-        //   limitation of the current implementation is that this
-        //   flag is not tracked per-environment or per-role.)  This
-        //   option can be used to resolve circular dependencies in
-        //   exceptional circumstances, eg, the 'meteor' package
-        //   depends on 'handlebars', but all packages (including
-        //   'handlebars') have an implicit dependency on
-        //   'meteor'. Internal use only -- future support of this
-        //   is not guaranteed. #UnorderedPackageReferences
-        //
-        // - weak: if true, don't require this package to load at all, but if
-        //   it's going to load, load it before us.  Don't bring this
-        //   package's imports into our namespace and don't allow us to use
-        //   its plugins. (Has the same limitation as "unordered" that this
-        //   flag is not tracked per-environment or per-role; this may
-        //   change.)
-
-        /**
-         * @memberOf PackageAPI
-         * @instance
-         * @summary Depend on package `packagename`.
-         * @locus package.js
-         * @param {String|String[]} packageNames Packages being depended on.
-         * Package names may be suffixed with an @version tag.
-         *
-         * In general, you must specify a package's version (e.g.,
-         * `'accounts@1.0.0'` to use version 1.0.0 or a higher
-         * compatible version (ex: 1.0.1, 1.5.0, etc.)  of the
-         * `accounts` package). If you are sourcing core
-         * packages from a Meteor release with `versionsFrom`, you may leave
-         * off version names for core packages. You may also specify constraints,
-         * such as `my:forms@=1.0.0` (this package demands `my:forms` at `1.0.0` exactly),
-         * or `my:forms@1.0.0 || =2.0.1` (`my:forms` at `1.x.y`, or exactly `2.0.1`).
-         * @param {String} [architecture] If you only use the package on the
-         * server (or the client), you can pass in the second argument (e.g.,
-         * `'server'`, `'client'`, `'web.browser'`, `'web.cordova'`) to specify
-         * what architecture the package is used with.
-         * @param {Object} [options]
-         * @param {Boolean} options.weak Establish a weak dependency on a
-         * package. If package A has a weak dependency on package B, it means
-         * that including A in an app does not force B to be included too — but,
-         * if B is included or by another package, then B will load before A.
-         * You can use this to make packages that optionally integrate with or
-         * enhance other packages if those packages are present.
-         * When you weakly depend on a package you don't see its exports.
-         * You can detect if the possibly-present weakly-depended-on package
-         * is there by seeing if `Package.foo` exists, and get its exports
-         * from the same place.
-         * @param {Boolean} options.unordered It's okay to load this dependency
-         * after your package. (In general, dependencies specified by `api.use`
-         * are loaded before your package.) You can use this option to break
-         * circular dependencies.
-         */
-        use: function (names, arch, options) {
-          // Support `api.use(package, {weak: true})` without arch.
-          if (_.isObject(arch) && !_.isArray(arch) && !options) {
-            options = arch;
-            arch = null;
-          }
-          options = options || {};
-
-          names = toArray(names);
-          arch = toArchArray(arch);
-
-          // A normal dependency creates an ordering constraint and a "if I'm
-          // used, use that" constraint. Unordered dependencies lack the
-          // former; weak dependencies lack the latter. There's no point to a
-          // dependency that lacks both!
-          if (options.unordered && options.weak) {
-            buildmessage.error(
-              "A dependency may not be both unordered and weak.",
-              { useMyCaller: true });
-            // recover by ignoring
-            return;
-          }
-
-          // using for loop rather than underscore to help with useMyCaller
-          for (var i = 0; i < names.length; ++i) {
-            var name = names[i];
-            try {
-              var parsed = utils.parsePackageConstraint(name);
-            } catch (e) {
-              if (!e.versionParserError)
-                throw e;
-              buildmessage.error(e.message, {useMyCaller: true});
-              // recover by ignoring
-              continue;
-            }
-
-            forAllMatchingArchs(arch, function (a) {
-              uses[a].push({
-                package: parsed.package,
-                constraint: parsed.constraintString,
-                unordered: options.unordered || false,
-                weak: options.weak || false
-              });
-            });
-          }
-        },
-
-        // Called when this package wants packages using it to also use
-        // another package.  eg, for umbrella packages which want packages
-        // using them to also get symbols or plugins from their components.
-
-        /**
-         * @memberOf PackageAPI
-         * @summary Give users of this package access to another package (by passing  in the string `packagename`) or a collection of packages (by passing in an  array of strings [`packagename1`, `packagename2`]
-         * @locus package.js
-         * @instance
-         * @param {String|String[]} packageSpecs Name of a package, or array of package names, with an optional @version component for each.
-         */
-        imply: function (names, arch) {
-          // We currently disallow build plugins in debugOnly packages; but if
-          // you could use imply in a debugOnly package, you could pull in the
-          // build plugin from an implied package, which would have the same
-          // problem as allowing build plugins directly in the package. So no
-          // imply either!
-          if (self.debugOnly) {
-            buildmessage.error("can't use imply in debugOnly packages");
-            // recover by ignoring
-            return;
-          }
-
-          names = toArray(names);
-          arch = toArchArray(arch);
-
-          // using for loop rather than underscore to help with useMyCaller
-          for (var i = 0; i < names.length; ++i) {
-            var name = names[i];
-            try {
-              var parsed = utils.parsePackageConstraint(name);
-            } catch (e) {
-              if (!e.versionParserError)
-                throw e;
-              buildmessage.error(e.message, {useMyCaller: true});
-              // recover by ignoring
-              continue;
-            }
-
-            forAllMatchingArchs(arch, function (a) {
-              // We don't allow weak or unordered implies, since the main
-              // purpose of imply is to provide imports and plugins.
-              implies[a].push({
-                package: parsed.package,
-                constraint: parsed.constraintString
-              });
-            });
-          }
-        },
-
-        // Top-level call to add a source file to a package. It will
-        // be processed according to its extension (eg, *.coffee
-        // files will be compiled to JavaScript).
-
-        /**
-         * @memberOf PackageAPI
-         * @instance
-         * @summary Specify the source code for your package.
-         * @locus package.js
-         * @param {String|String[]} filename Name of the source file, or array of strings of source file names.
-         * @param {String} [architecture] If you only want to export the file
-         * on the server (or the client), you can pass in the second argument
-         * (e.g., 'server', 'client', 'web.browser', 'web.cordova') to specify
-         * what architecture the file is used with.
-         */
-        addFiles: function (paths, arch, fileOptions) {
-          paths = toArray(paths);
-          arch = toArchArray(arch);
-
-          // Convert Dos-style paths to Unix-style paths.
-          // XXX it is possible to convert an already Unix-style path by mistake
-          // and break it. e.g.: 'some\folder/anotherFolder' is a valid path
-          // consisting of two components. #WindowsPathApi
-          paths = _.map(paths, function (p) {
-            if (p.indexOf('/') !== -1) {
-              // it is already a Unix-style path most likely
-              return p;
-            }
-            return files.convertToPosixPath(p, true);
-          });
-
-          _.each(paths, function (path) {
-            forAllMatchingArchs(arch, function (a) {
-              var source = {relPath: path};
-              if (fileOptions)
-                source.fileOptions = fileOptions;
-              sources[a].push(source);
-            });
-          });
-        },
-
-        // Use this release to resolve unclear dependencies for this package. If
-        // you don't fill in dependencies for some of your implies/uses, we will
-        // look at the packages listed in the release to figure that out.
-
-        /**
-         * @memberOf PackageAPI
-         * @instance
-         * @summary Use versions of core packages from a release. Unless provided, all packages will default to the versions released along with `meteorRelease`. This will save you from having to figure out the exact versions of the core packages you want to use. For example, if the newest release of meteor is `METEOR@0.9.0` and it includes `jquery@1.0.0`, you can write `api.versionsFrom('METEOR@0.9.0')` in your package, and when you later write `api.use('jquery')`, it will be equivalent to `api.use('jquery@1.0.0')`. You may specify an array of multiple releases, in which case the default value for constraints will be the "or" of the versions from each release: `api.versionsFrom(['METEOR@0.9.0', 'METEOR@0.9.5'])` may cause `api.use('jquery')` to be interpreted as `api.use('jquery@1.0.0 || 2.0.0')`.
-         * @locus package.js
-         * @param {String | String[]} meteorRelease Specification of a release: track@version. Just 'version' (e.g. `"0.9.0"`) is sufficient if using the default release track `METEOR`.
-         */
-        versionsFrom: function (releases) {
-          // Packages in isopackets really ought to be in the core release, by
-          // definition, so saying that they should use versions from another
-          // release doesn't make sense. Moreover, if we're running from a
-          // checkout, we build isopackets before we initialize catalog.official
-          // (since we may need the ddp isopacket to refresh catalog.official),
-          // so we wouldn't actually be able to interpret the release name
-          // anyway.
-          if (initFromPackageDirOptions.buildingIsopackets) {
-            buildmessage.error(
-              "packages in isopackets may not use versionsFrom");
-            // recover by ignoring
-            return;
-          }
-
-          releases = toArray(releases);
-
-          // using for loop rather than underscore to help with useMyCaller
-          for (var i = 0; i < releases.length; ++i) {
-            var release = releases[i];
-
-            // If you don't specify a track, use our default.
-            if (release.indexOf('@') === -1) {
-              release = catalog.DEFAULT_TRACK + "@" + release;
-            }
-
-            var relInf = release.split('@');
-            if (relInf.length !== 2) {
-              buildmessage.error("Release names in versionsFrom may not contain '@'.",
-                                 { useMyCaller: true });
-              return;
-            }
-            var releaseRecord = catalog.official.getReleaseVersion(
-              relInf[0], relInf[1]);
-            if (!releaseRecord) {
-              buildmessage.error("Unknown release "+ release,
-                                 { tags: { refreshCouldHelp: true } });
-            } else {
-              releaseRecords.push(releaseRecord);
-            }
-          }
-        },
-
-        // Export symbols from this package.
-        //
-        // @param symbols String (eg "Foo") or array of String
-        // @param arch 'web', 'server', 'web.browser', 'web.cordova'
-        // or an array of those.
-        // The default is ['web', 'server'].
-        // @param options 'testOnly', boolean.
-
-        /**
-         * @memberOf PackageAPI
-         * @instance
-         * @summary Export package-level variables in your package. The specified variables (declared without `var` in the source code) will be available to packages that use this package.
-         * @locus package.js
-         * @param {String} exportedObject Name of the object.
-         * @param {String} [architecture] If you only want to export the object
-         * on the server (or the client), you can pass in the second argument
-         * (e.g., 'server', 'client', 'web.browser', 'web.cordova') to specify
-         * what architecture the export is used with.
-         */
-        export: function (symbols, arch, options) {
-          // Support `api.export("FooTest", {testOnly: true})` without
-          // arch.
-          if (_.isObject(arch) && !_.isArray(arch) && !options) {
-            options = arch;
-            arch = null;
-          }
-          options = options || {};
-
-          symbols = toArray(symbols);
-          arch = toArchArray(arch);
-
-          _.each(symbols, function (symbol) {
-            // XXX be unicode-friendlier
-            if (!symbol.match(/^([_$a-zA-Z][_$a-zA-Z0-9]*)$/)) {
-              buildmessage.error("Bad exported symbol: " + symbol,
-                                 { useMyCaller: true });
-              // recover by ignoring
-              return;
-            }
-            forAllMatchingArchs(arch, function (w) {
-              exports[w].push({name: symbol, testOnly: !!options.testOnly});
-            });
-          });
-        }
-      };
-
-      // XXX COMPAT WITH 0.8.x
-      api.add_files = api.addFiles;
-
       try {
         buildmessage.markBoundary(fileAndDepLoader)(api);
       } catch (e) {
@@ -1354,9 +972,9 @@ _.extend(PackageSource.prototype, {
         // packages and any remaining handlers. It violates the
         // principle of least surprise to half-run a handler
         // and then continue.
-        sources = {};
+        api.sources = {};
         _.each(compiler.ALL_ARCHES, function (arch) {
-          sources[arch] = [];
+          api.sources[arch] = [];
         });
 
         fileAndDepLoader = null;
@@ -1375,14 +993,14 @@ _.extend(PackageSource.prototype, {
       }
     };
     _.each(compiler.ALL_ARCHES, function (label) {
-      _.each(uses[label], doNotDepOnSelf);
-      _.each(implies[label], doNotDepOnSelf);
+      _.each(api.uses[label], doNotDepOnSelf);
+      _.each(api.implies[label], doNotDepOnSelf);
     });
 
     // If we have specified some release, then we should go through the
     // dependencies and fill in the unspecified constraints with the versions in
     // the releases (if possible).
-    if (!_.isEmpty(releaseRecords)) {
+    if (!_.isEmpty(api.releaseRecords)) {
 
       // Given a dependency object with keys package (the name of the package)
       // and constraint (the version constraint), if the constraint is null,
@@ -1393,7 +1011,7 @@ _.extend(PackageSource.prototype, {
           return dep;
         }
         var newConstraint = [];
-        _.each(releaseRecords, function (releaseRecord) {
+        _.each(api.releaseRecords, function (releaseRecord) {
           var packages = releaseRecord.packages;
           if(_.has(packages, dep.package)) {
             newConstraint.push(packages[dep.package]);
@@ -1407,11 +1025,11 @@ _.extend(PackageSource.prototype, {
         return dep;
       };
 
-      // For all implies and uses, fill in the unspecified dependencies from the
+      // For all api.implies and api.uses, fill in the unspecified dependencies from the
       // release.
       _.each(compiler.ALL_ARCHES, function (label) {
-        uses[label] = _.map(uses[label], setFromRel);
-        implies[label] = _.map(implies[label], setFromRel);
+        api.uses[label] = _.map(api.uses[label], setFromRel);
+        api.implies[label] = _.map(api.implies[label], setFromRel);
       });
      };
 
@@ -1466,11 +1084,11 @@ _.extend(PackageSource.prototype, {
         // dependency on meteor dating from when the .js extension handler was
         // in the "meteor" package).
         var alreadyDependsOnMeteor =
-              !! _.find(uses[arch], function (u) {
+              !! _.find(api.uses[arch], function (u) {
                 return u.package === "meteor";
               });
         if (! alreadyDependsOnMeteor)
-          uses[arch].unshift({ package: "meteor" });
+          api.uses[arch].unshift({ package: "meteor" });
       }
 
       // Each unibuild has its own separate WatchSet. This is so that, eg, a test
@@ -1483,10 +1101,10 @@ _.extend(PackageSource.prototype, {
       self.architectures.push(new SourceArch(self, {
         kind: "main",
         arch: arch,
-        uses: uses[arch],
-        implies: implies[arch],
-        getSourcesFunc: function () { return sources[arch]; },
-        declaredExports: exports[arch],
+        uses: api.uses[arch],
+        implies: api.implies[arch],
+        getSourcesFunc: function () { return api.sources[arch]; },
+        declaredExports: api.exports[arch],
         watchSet: watchSet
       }));
     });
